refactored contact form styles
