+++ src/App.jsx
@@ -2,7 +2,7 @@ import React, { useEffect } from "react";
 import Header from "./components/Header/Header";
 import Hero from "./components/Hero/Hero";
 import Menu from "./components/Menu/Menu";
-import Access from "./components/Access";
+import Access from "./components/Access/Access";
 import Footer from "./components/Footer/Footer";
 import { initializeScripts } from "./assets/js/main";
 import "./App.css";
